+++ doc/changelog.md
@@ -8,6 +8,7 @@ All notable changes to Mili-python will be documented in this file.
 - Added the function `MiliDatabase.state_count` to get the number of states in a database.
 
 ### Fixed
+- Fixed bugs related to querying array state variables
 
 ### Changed